+++ src/model/custom-error.ts
